신용도검산, 신용도 로고 제작 및 중기업체 Repository, leaseService 내에 중기업체 조회에 관련된 Service메서드 추가

+++ leaseCity/src/main/java/leasecity/dto/user/Credit.java
@@ -21,5 +21,5 @@ public class Credit implements Serializable {
 	private String acceptUserId;
 
 	/** 평가점수. */
-	private Float evaluation;
+	private Double evaluation;
 }
